pyproject.toml
[tool.poetry]
name = "sisyphus-matroska"
version = "0.9.0"
description = ""
authors = ["JamesTheBard <jamesthebard@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
packages = [
    { include = "mkvmerge.py" },
    { include = "mkvextract.py" },
    { include = "mkvinfo.py" }
]
include = [
    "schema/mimetypes.schema.json",
    "schema/mkvextract.schema.json",
    "schema/mkvmerge.schema.json"
]

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.5.2"
python-box = "^7.0.1"
jsonschema = "^4.19.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
